microservicio/dominio/src/test/java/com/ceiba/producto/servicio/ServicioCrearProductoTest.java
package com.ceiba.producto.servicio;

import com.ceiba.BasePrueba;
import com.ceiba.dominio.excepcion.ExcepcionDuplicidad;
import com.ceiba.producto.modelo.entidad.Producto;
import com.ceiba.producto.puerto.repositorio.RepositorioProducto;
import com.ceiba.producto.servicio.testdatabuilder.ProductoTestDataBuilder;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.mockito.Mockito;

public class ServicioCrearProductoTest {

    @Test
    @DisplayName("Deberia lanzar una exepcion cuando se valide la existencia del Producto")
    void deberiaLanzarUnaExepcionCuandoSeValideLaExistenciaDelProducto() {
        // arrange
        Producto producto = new ProductoTestDataBuilder().build();
        RepositorioProducto repositorioProducto = Mockito.mock(RepositorioProducto.class);
        Mockito.when(repositorioProducto.existe(Mockito.anyString())).thenReturn(true);
        ServicioCrearProducto servicioCrearProducto = new ServicioCrearProducto(repositorioProducto);
        // act - assert
        BasePrueba.assertThrows(() -> servicioCrearProducto.ejecutar(producto), ExcepcionDuplicidad.class,
                "El producto ya se encuentra registrado");
    }

    @Test
    @DisplayName("Deberia Crear el producto de manera correcta")
    void deberiaCrearElProductoDeManeraCorrecta() {
        // arrange
        Producto producto = new ProductoTestDataBuilder().build();
        RepositorioProducto repositorioProducto = Mockito.mock(RepositorioProducto.class);
        Mockito.when(repositorioProducto.existe(Mockito.anyString())).thenReturn(false);
        ServicioCrearProducto servicioCrearProducto = new ServicioCrearProducto(repositorioProducto);
        // act
        servicioCrearProducto.ejecutar(producto);
        // - assert
        Mockito.verify(repositorioProducto, Mockito.times(1)).crear(producto);
    }
}
